Refactor plugin management using Lua Module
