+++ README.md
@@ -5,17 +5,18 @@
 This is Neovim-Config Repository for my personal use.  
 (Since 2023/3/9~)
 
-This Repository has three Branches.  
-- main
+This Repository has three Branches.
+
+- `main`
   - Almost empty branch(only README.md exists)
-- coc
+- `coc`(current)
   - Neovim Configs based on Coc.nvim.(my daily-use settings)
-- lsp
+- `lsp`(not used)
   - Neovim Configs based on Neovim-Builtin LSP.
 
-You can choice favorite Branch(if you want to use)👐  
+You can choice favorite Branch(if you want to use)👐
 
-`$ git clone git@github.com:Nyowa450/nvim.git`
+`$ git clone git@github.com:KamabokoFish/nvim.git
 
 **This is my PERSONAL NEOVIM CONFIGS.**  
 **If you use this Configs (especially in practice), BE CAREFUL.🚨**
