+++ lua/plugins/standard.lua
@@ -1,0 +1,9 @@
+-- standard
+
+return {
+
+  {'nvim-lua/plenary.nvim'},
+
+  {'nvim-tree/nvim-web-devicons',lazy = false,priority = 1000},
+
+}
