--- plugin/colorizer.rc.lua
@@ -1,7 +1,0 @@
-local status,color = pcall(require,"colorizer")
-if (not status) then return end
-
-color.setup{
-  'css';
-  'scss';
-}
