--- plugin/git.rc.lua
@@ -1,14 +1,0 @@
---vim-fugitiveとNeogitを併用
-
-vim.cmd([[
-nnoremap <silent> <space>s :Git<CR><C-w>T
-nnoremap <silent> <space>b :Git blame<CR>
-nnoremap <space>d :Gvdiffsplit
-nnoremap <silent> <space>a :Gwrite<CR>
-]])
--- nnoremap <silent> <space>l :Git log --graph<CR><C-w>T
-
--- neogit
--- local neogit = require('neogit')
-
--- neogit.setup {}
