--- plugin/hlslens.rc.lua
@@ -1,21 +1,0 @@
-local status,hls = pcall(require,"hlslens")
-if (not status) then return end
-
-hls.setup({
-    calm_down = false,
-    -- nearest_only = true,
-})
-
--- local kopts = {noremap = true, silent = true}
-
--- vim.api.nvim_set_keymap('n', 'n',
---     [[<Cmd>execute('normal! ' . v:count1 . 'n')<CR><Cmd>lua require('hlslens').start()<CR>]],
---     kopts)
--- vim.api.nvim_set_keymap('n', 'N',
---     [[<Cmd>execute('normal! ' . v:count1 . 'N')<CR><Cmd>lua require('hlslens').start()<CR>]],
---     kopts)
--- vim.api.nvim_set_keymap('n', '*', [[*<Cmd>lua require('hlslens').start()<CR>]], kopts)
--- vim.api.nvim_set_keymap('n', '#', [[#<Cmd>lua require('hlslens').start()<CR>]], kopts)
--- vim.api.nvim_set_keymap('n', 'g*', [[g*<Cmd>lua require('hlslens').start()<CR>]], kopts)
--- vim.api.nvim_set_keymap('n', 'g#', [[g#<Cmd>lua require('hlslens').start()<CR>]], kopts)
-
