--- plugin/surround.rc.lua
@@ -1,6 +1,0 @@
-local status,surr = pcall(require,"nvim-surround")
-if (not status) then return end
-
-surr.setup({
-    -- Configuration here, or leave empty to use defaults
-})
